+++ Builds/changes.md
@@ -18,4 +18,6 @@
 - [x] Fix file size gathering function to be slightly more efficient
 - [x] Fix warp-cli version/size checking
 - [x] Fix tap icon/info button logic
+- [x] Fix homebrew auto-update not being able to create plist file if LaunchAgents folder is not present - #469
+- [x] Fix homebrew auto-update schedule hiccup with deleted schedules coming back from appstorage
 - [x] Translations